backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
paho-mqtt>=1.6
scikit-learn>=1.3
joblib>=1.3
numpy>=1.24
